% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBoundaries.R
\name{getBoundaries}
\alias{getBoundaries}
\title{Determine the Boundaries of a Plot in Terms of the Data Units}
\usage{
getBoundaries(region)
}
\arguments{
\item{region}{The region of the plot to use for defining the boundaries. Must
be one of "device", "figure", "plot", or "data".}
}
\value{
A numeric vector of length four giving the coordinates of the plotting
        boundary, in the order of bottom, left, top, right.
}
\description{
Determine the Boundaries of a Plot in Terms of the Data Units
}
\examples{
\dontrun{

plot(1:10)
print(getBoundaries('data'))
print(getBoundaries('plot'))
print(getBoundaries('figure'))
print(getBoundaries('device'))

}

}
\author{
Jasper Watson
}
