% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relativeUnitConversions.R
\name{getDataPerInch}
\alias{getDataPerInch}
\title{Find the Number of Data Points Per Inch in a Plot}
\usage{
getDataPerInch()
}
\value{
Numeric vector of length two giving the number of data points per inch,
        for the x axis and y axis, respectively.
}
\description{
Find the Number of Data Points Per Inch in a Plot
}
\author{
Jasper Watson
}
