% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relativeUnitConversions.R
\name{getDataPerPixel}
\alias{getDataPerPixel}
\title{Find the Number of Data Points Per Pixel in a Plot}
\usage{
getDataPerPixel()
}
\value{
Numeric vector of length two giving the number of data points per pixel,
        for the x axis and y axis, respectively.
}
\description{
Find the Number of Data Points Per Pixel in a Plot
}
\examples{
\dontrun{

plot(1:10)
print(getDataPerPixel())

}

}
\author{
Jasper Watson
}
