% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relativeUnitConversions.R
\name{getInchesPerLine}
\alias{getInchesPerLine}
\title{Find the Number of Inches Per Margin Line in a Plot}
\usage{
getInchesPerLine()
}
\value{
Numeric vector of length two giving the number of inches per line,
        for the x axis and y axis, respectively.
}
\description{
Find the Number of Inches Per Margin Line in a Plot
}
\note{
The number of lines per inch is the same for both the x and y axes but
      we return a vector of length two to maintain consistency with all of the
      other getXperY style functions.
}
\examples{
\dontrun{

plot(1:10)
print(getInchesPerLine())

}

}
\author{
Jasper Watson
}
