% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relativeUnitConversions.R
\name{getInchesPerPixel}
\alias{getInchesPerPixel}
\title{Find the Number of Inches Per Pixel in a Plot}
\usage{
getInchesPerPixel()
}
\value{
Numeric vector of length two giving the number of inches per pixel,
        for the x axis and y axis, respectively.
}
\description{
Find the Number of Inches Per Pixel in a Plot
}
\examples{
\dontrun{

plot(1:10)
print(getInchesPerPixel())

}

}
\author{
Jasper Watson
}
