% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relativeUnitConversions.R
\name{getLinesPerDatum}
\alias{getLinesPerDatum}
\title{Find the Number of Margin Lines Per Data Point in a Plot}
\usage{
getLinesPerDatum()
}
\value{
Numeric vector of length two giving the number of lines per data point,
for the x axis and y axis, respectively.
}
\description{
Find the Number of Margin Lines Per Data Point in a Plot
}
\examples{
\dontrun{

plot(1:10)
print(getLinesPerDatum())

}

}
\author{
Jasper Watson
}
