% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relativeUnitConversions.R
\name{getLinesPerPixel}
\alias{getLinesPerPixel}
\title{Find the Number of Margin Lines Per Pixel in a Plot}
\usage{
getLinesPerPixel()
}
\value{
Numeric vector of length two giving the number of lines per pixel,
        for the x axis and y axis, respectively.
}
\description{
Find the Number of Margin Lines Per Pixel in a Plot
}
\examples{
\dontrun{

plot(1:10)
print(getLinesPerPixel())

}

}
\author{
Jasper Watson
}
