% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relativeUnitConversions.R
\name{getPixelsPerDatum}
\alias{getPixelsPerDatum}
\title{Find the Number of Pixels Per Data Point in a Plot}
\usage{
getPixelsPerDatum()
}
\value{
Numeric vector of length two giving the number of pixels per data point,
for the x axis and y axis, respectively.
}
\description{
Find the Number of Pixels Per Data Point in a Plot
}
\examples{
\dontrun{

plot(1:10)
print(getPixelsPerDatum())

}

}
\author{
Jasper Watson
}
