% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relativeUnitConversions.R
\name{getPixelsPerInch}
\alias{getPixelsPerInch}
\title{Find the Number of Pixels Per Inch in a Plot}
\usage{
getPixelsPerInch()
}
\value{
Numeric vector of length two giving the number of pixels per inch,
        for the x axis and y axis, respectively.
}
\description{
Find the Number of Pixels Per Inch in a Plot
}
\examples{
\dontrun{

plot(1:10)
print(getPixelsPerInch())

}

}
\author{
Jasper Watson
}
