% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relativeUnitConversions.R
\name{getPixelsPerLine}
\alias{getPixelsPerLine}
\title{Find the Number of Pixels Per Margin Line in a Plot}
\usage{
getPixelsPerLine()
}
\value{
Numeric vector of length two giving the number of pixels per line,
for the x axis and y axis, respectively.
}
\description{
Find the Number of Pixels Per Margin Line in a Plot
}
\examples{
\dontrun{

plot(1:10)
print(getPixelsPerLine())

}

}
\author{
Jasper Watson
}
