% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotatePlot.R
\name{highlightDeviceRegion}
\alias{highlightDeviceRegion}
\title{Highlight the Device Region of a Plot}
\usage{
highlightDeviceRegion(border = "skyblue", col = adjustcolor(border, 0.1))
}
\arguments{
\item{border}{Parameter to be passed to graphics::rect. Defaults to "green".}

\item{col}{Parameter to be passed to graphics::rect. Defaults to
adjustcolor(border, 0.1).}
}
\description{
Highlight the Device Region of a Plot
}
\examples{
\dontrun{

plot(1:10)
par(xpd = NA)
highlightDeviceRegion()

}

}
\author{
Jasper Watson
}
