% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineLocations.R
\name{lineLocations}
\alias{lineLocations}
\title{Wrapper Function for convertUnits Focused on Identifying the Locations of
the Margin Lines of a Plot}
\usage{
lineLocations(side, line)
}
\arguments{
\item{side}{Integer giving the side of the plot to count lines from.}

\item{line}{Numeric vector giving margin lines one wishes to find the data
coordinates of.}
}
\description{
Wrapper Function for convertUnits Focused on Identifying the Locations of
the Margin Lines of a Plot
}
\note{
No attempt is made to limit the returned values to the device region.
}
\examples{
\dontrun{

## Illustrate where the lines fall when using mtext:
plot(1:10)
mtext(1:26, line = -(1:26), side = 1, col = 'blue')
x <- lineLocations(1, 0:(-26))
abline(h = x, col = 'red', lty = 2)
}

}
\author{
Jasper Watson
}
