% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precisePlacement.R
\docType{package}
\name{precisePlacement}
\alias{precisePlacement}
\title{Suite of Functions To Get Plot Elements Exactly Where You Want Them}
\description{
This package provides a selection of tools that make it easier to place
elements onto a (base R) plot exactly where you want them. It allows users
to identify points and distances on a plot in terms of inches, pixels,
margin lines, data units, and proportions of the plotting space, all in a
manner more simple than manipulating par(). It helps users identify where
exactly the margin lines of a plot fall and to create sub-figures.
}
\author{
Jasper Watson
}
