% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showMarginLines.R
\name{showMarginLines}
\alias{showMarginLines}
\title{Add Lines to a Plot to Indicate Where the Margins Are}
\usage{
showMarginLines(sides = 1:4, lty = 2, ...)
}
\arguments{
\item{sides}{Numeric vector giving the sides of the plot to annotate. Defaults
to 1:4.}

\item{lty}{Argument to be passed to graphics::abline. Defaults to 2.}

\item{...}{Other arguments to be passed to graphics::abline}
}
\description{
Add Lines to a Plot to Indicate Where the Margins Are
}
\examples{
\dontrun{

plot(1:10)
par(xpd = NA)
showMarginLines()

}

}
\author{
Jasper Watson
}
