% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showOuterMarginLines.R
\name{showOuterMarginLines}
\alias{showOuterMarginLines}
\title{Add Lines to a Plot to Indicate Where the Outer Margins Are}
\usage{
showOuterMarginLines(sides = 1:4, lty = 3, col = "purple", ...)
}
\arguments{
\item{sides}{Numeric vector giving the sides of the plot to annotate. Defaults
to 1:4.}

\item{lty}{Argument to be passed to graphics::abline. Defaults to 3.}

\item{col}{Argument to be passed to graphics::abline. Defaults to "purple".}

\item{...}{Other arguments to be passed to graphics::abline}
}
\description{
Add Lines to a Plot to Indicate Where the Outer Margins Are
}
\examples{
\dontrun{

par(oma = 1:4, mfrow = 2:1)
plot(1:10)
plot(1:10)
par(xpd = NA)
showOuterMarginLines()

}

}
\author{
Jasper Watson
}
